+++ myems-admin/app/controllers/settings/equipment/equipment.controller.js
@@ -12,7 +12,8 @@ app.controller('EquipmentController', function(
 	SweetAlert) {
 	$scope.cur_user = JSON.parse($window.localStorage.getItem("myems_admin_ui_current_user"));
 	$scope.getAllEquipments = function() {
-		EquipmentService.getAllEquipments(function (response) {
+		let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+		EquipmentService.getAllEquipments(headers, function (response) {
 			if (angular.isDefined(response.status) && response.status === 200) {
 				$scope.equipments = response.data;
 			} else {