+++ myems-api/reports/equipmentstatistics.py
@@ -10,6 +10,7 @@ import simplejson as json
 import config
 import excelexporters.equipmentstatistics
 from core import utilities
+from core.useractivity import access_control
 
 
 class Reporting:
@@ -36,6 +37,7 @@ class Reporting:
     ####################################################################################################################
     @staticmethod
     def on_get(req, resp):
+        access_control(req)
         print(req.params)
         equipment_id = req.params.get('equipmentid')
         equipment_uuid = req.params.get('equipmentuuid')