+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - added access control to tenant in myems-api, myems-admin and myems-web
 - added point real time report to myems-api
+- added access control to equipment in myems-api, myems-admin and myems-web
 - 
 ### Changed
 -